Moving parts into ui directory.
