+++ src/ts/editor/ui/parts/content/footer.ts
