+++ src/ts/editor/ui/parts/menu/workspaces.ts
@@ -1,17 +1,17 @@
-import {ApiError, WorkspaceData} from '../../api';
+import {ApiError, WorkspaceData} from '../../../api';
 import {
   DeepObject,
   TemplateResult,
   classMap,
   html,
 } from '@blinkk/selective-edit';
-import {DialogActionLevel, FormDialogModal} from '../../ui/modal';
+import {DialogActionLevel, FormDialogModal} from '../../modal';
 import {MenuSectionPart, MenuSectionPartConfig} from './index';
-import {EVENT_WORKSPACE_LOAD} from '../../events';
-import {FeatureFlags} from '../../features';
+import {EVENT_WORKSPACE_LOAD} from '../../../events';
+import {FeatureFlags} from '../../../features';
 import merge from 'lodash.merge';
 import {repeat} from '@blinkk/selective-edit';
-import {templateLoading} from '../../template';
+import {templateLoading} from '../../../template';
 
 const MODAL_KEY_NEW = 'menu_workspace_new';
